Nice.
Currently if the AST gets too big it goes off the right side of the screen and I cannot scroll, same for the tromp diagram.

For the AST, can we make sure that the window there is scrollable, but also make it much more vertical in how it reads so that it (in general) fits into the x-scale of the box. Similarly for the AST I don't like that I'm manually controlling the scale, can it just pick the scale to make sure it fits on the screen