Absolutely beautiful. Amazing work.
Let's go back to tromp diagrams. So in a Tromp diagram the bottom bar represents application so you can see that the "right" block gets "copied" into the left block (this is application). I want to animate this process so the bottom application "bar" contracts right to left while still touching the whole right block, and then the right block shrinks/expands to the new size it needs to and gets "absorbed" into the left block on each update